2n_Gen_Cloud_Functions/add-activity/requirements.txt
functions-framework==3.*
Flask==2.3.3
google-cloud-firestore==2.16.0
google-generativeai==0.5.4
assemblyai